dodanie podstawowych funkcji matematycznych

+++ app/src/main/java/pl/angrymarschmallow/calculatoran/CalculatorEngine.java
@@ -9,7 +9,7 @@ public interface CalculatorEngine {
     double sum(double... a);             // Dodawanie
     double substr(double a, double b);   // Odejmowanie
     double pow(double a, double b);      // Potęgowanie
-    double sqrt(double a, double b);     // Pierwiastkowanie
+    double sqrt(double a, int b);     // Pierwiastkowanie
 
     double[] quadratic(double a, double b, double c); // Zakładam, że zwraca listę pierwiastków i deltę, opisz dokładnie
